requirements.txt
absl-py==0.7.1
appnope==0.1.0
argcomplete==1.10.0
asn1crypto==0.24.0
astor==0.8.0
backcall==0.1.0
bleach==2.1.4
boto==2.49.0
certifi==2016.9.26
cffi==1.12.3
Click==7.0
colorama==0.4.1
crcmod==1.7
cryptography==2.7
cycler==0.10.0
Cython==0.28.5
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.2.3
enum34==1.1.6
fasteners==0.15
gast==0.2.2
gcs-oauth2-boto-plugin==2.5
google-apitools==0.5.30
google-pasta==0.1.7
google-reauth==0.1.0
grpcio==1.22.0
gsutil==4.40
h5py==2.9.0
html5lib==1.0.1
httplib2==0.13.0
ipykernel==4.10.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.0
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.5.0
jupyterlab==0.34.9
jupyterlab-launcher==0.13.1
jupyterthemes==0.20.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-tqdm==2.0.1
kiwisolver==1.0.1
lesscpy==0.13.0
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.0.0
mistune==0.8.3
mkl-fft==1.0.6
mkl-random==1.0.1
mock==2.0.0
monotonic==1.5
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.6.0
numpy==1.15.2
oauth2client==4.1.3
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
pbr==5.4.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==6.1.0
plaidbench==0.6.2
plaidml==0.6.2
plaidml-keras==0.6.2
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==1.0.15
protobuf==3.8.0
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.0
python-dateutil==2.7.3
pytz==2019.1
pyu2f==0.1.4
PyYAML==5.1.1
pyzmq==17.1.2
retry-decorator==1.1.0
rsa==4.0
scipy==1.3.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.12.0
SocksiPy-branch==1.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.1.1
tqdm==4.32.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
wrapt==1.11.2
